JavaScript clean up: .text(), .html(), !(>), closure.
.html() uses innerHTML which makes the browser parse the text as
HTML. Though the input is safe in this case and doesn't contain
special characters that need escaping, it is a bad habit and is
slightly slower then just inserting a the text directly in the
dom as text. Using .text() instead.

Also fixes a bad call to .html() in fh.resetForm (calling .html()
is a getter, returns the html, which is not intended here. The
likely intention here is to empty it). To do that use html('') or
simply .empty() (which is what .html() will call internally).

Converted !(>) to <= (lint warning).

/index.html:
Move functions outside the dom-ready handler, improves performance
slightly (due to block level change, best reviewed with ?w=1 / -w)